simplify statemgmt, put provider at top of treecleaned up state mgmt after move to Provider and moved musica data providerup to the top of the tree to be available for all routes
