+++ lib/main.dart
@@ -1,4 +1,7 @@
+import 'package:flute_music_player/flute_music_player.dart';
 import 'package:flutter/material.dart';
+import 'package:music_home/data/music_data.dart';
+import 'package:provider/provider.dart';
 import 'my_app.dart';
 import 'utils/themes.dart';
 
@@ -14,10 +17,13 @@ class MyMaterialApp extends StatefulWidget {
 class MyMaterialAppState extends State<MyMaterialApp> {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      theme: darkTheme,
-      home: MyApp(),
+    return ChangeNotifierProvider<MusicData>(
+      create: (_) => MusicData(MusicFinder()),
+      child: MaterialApp(
+        debugShowCheckedModeBanner: false,
+        theme: darkTheme,
+        home: MyApp(),
+      ),
     );
   }
 }
