+++ lib/my_app.dart
@@ -1,9 +1,7 @@
-import 'package:flute_music_player/flute_music_player.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:provider/provider.dart';
 
-import 'data/song_data.dart';
+import 'data/music_data.dart';
 import 'pages/root_page.dart';
 
 class MyApp extends StatefulWidget {
@@ -25,8 +23,6 @@ class _MyAppState extends State<MyApp> {
       DeviceOrientation.portraitUp,
     ]);
     SystemChrome.setEnabledSystemUIOverlays([]);
-
-    initPlatformState();
   }
 
   @override
@@ -35,33 +31,8 @@ class _MyAppState extends State<MyApp> {
     songData.audioPlayer.stop();
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
-  void initPlatformState() async {
-    List<Track> songs;
-    try {
-      songs = (await MusicFinder.allSongs() as List<Song>)
-          .map((s) => Track.fromSong(s))
-          .toList();
-    } catch (e) {
-      print("Failed to get songs: '$e'.");
-    }
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) {
-      return;
-    }
-
-    setState(() {
-      songData = MusicData(songs, MusicFinder());
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
-    return Provider<MusicData>.value(
-      value: songData,
-      child: RootPage(),
-    );
+    return RootPage();
   }
 }
