+++ lib/pages/album_tracklist.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:music_home/widgets/now_playing_button.dart';
-import '../data/song_data.dart';
+import '../data/music_data.dart';
 import '../widgets/mp_listview.dart';
 
 class AlbumTrackList extends StatelessWidget {
@@ -15,7 +15,7 @@ class AlbumTrackList extends StatelessWidget {
       appBar: AppBar(
         title: Text(album.title),
         actions: <Widget>[
-          //NowPlayingButton(),
+          NowPlayingButton(),
         ],
       ),
       body: MPListView(data, false, tracks: album.tracks),
