+++ lib/widgets/now_playing_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:music_home/data/song_data.dart';
+import 'package:music_home/data/music_data.dart';
 import 'package:music_home/pages/now_playing.dart';
 import 'package:provider/provider.dart';
